# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CLinic Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/medical/cases": {
            "post": {
                "tags": [
                    "Medical Cases"
                ],
                "summary": "Create a new medical case",
                "description": "Store a newly created medical case in storage.",
                "operationId": "e8466fd43780d0d7f923b4743b2bfbaa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMedicalCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Medical case created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicalCaseResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/medical-cases/{id}": {
            "put": {
                "tags": [
                    "Medical Cases"
                ],
                "summary": "Update a specific medical case",
                "description": "Update the specified medical case in storage.",
                "operationId": "da184fdc279bd411ba4e6927c7c627bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the medical case to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMedicalCaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Medical case updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicalCaseResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Medical Cases"
                ],
                "summary": "Delete a specific medical case",
                "description": "Remove the specified medical case from storage.",
                "operationId": "77eafcf81316938c01a9935f800bde66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the medical case to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Medical case deleted"
                    }
                }
            }
        },
        "/api/medical/sessions": {
            "post": {
                "tags": [
                    "Medical Sessions"
                ],
                "summary": "Create a new medical session",
                "operationId": "6f12bffb96fb98ba4335c5fb0cdeb04c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicalSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Medical session created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicalSessionResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/medical/sessions/{medicalSession}": {
            "put": {
                "tags": [
                    "Medical Sessions"
                ],
                "summary": "Update a specific medical session",
                "operationId": "98bd053fcd31bd7d383215f914c270c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the medical session to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MedicalSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Medical session updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MedicalSessionResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Medical Sessions"
                ],
                "summary": "Delete a specific medical session",
                "operationId": "44102222057548b9a3084520c628c844",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the medical session to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Medical session deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MedicalSessionRequest": {
                "title": "MedicalSessionRequest",
                "description": "Request schema for creating or updating a medical session",
                "required": [
                    "medicalCaseId",
                    "date"
                ],
                "properties": {
                    "medicalCaseId": {
                        "description": "The ID of the medical case",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "date": {
                        "description": "The date of the medical session",
                        "type": "string",
                        "format": "date",
                        "example": "2023-10-01"
                    }
                },
                "type": "object"
            },
...
            "StoreMedicalCaseRequest": {
                "title": "Store Medical Case Request",
                "description": "Request body for storing a medical case",
                "required": [
                    "name",
                    "clinic_id",
                    "total",
                    "date",
                    "patientId"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the medical case",
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "description": "The description of the medical case",
                        "type": "string"
                    },
                    "clinic_id": {
                        "description": "The ID of the clinic associated with the medical case",
                        "type": "integer"
                    },
                    "total": {
                        "description": "The total cost of the medical case",
                        "type": "number",
                        "format": "float"
                    },
                    "date": {
                        "description": "The date of the medical case",
                        "type": "string",
                        "format": "date"
                    },
                    "patientId": {
                        "description": "The ID of the patient associated with the medical case",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
            "UpdateMedicalCaseRequest": {
                "title": "Update Medical Case Request",
                "description": "Request body for updating a medical case",
                "required": [],
                "properties": {
                    "name": {
                        "description": "The name of the medical case",
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "description": "The description of the medical case",
                        "type": "string"
                    },
                    "clinic_id": {
                        "description": "The ID of the clinic associated with the medical case",
                        "type": "integer"
                    },
                    "total": {
                        "description": "The total cost of the medical case",
                        "type": "number",
                        "format": "float"
                    },
                    "date": {
                        "description": "The date of the medical case",
                        "type": "string",
                        "format": "date"
                    },
                    "patientId": {
                        "description": "The ID of the patient associated with the medical case",
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "type": "object"
            },
...
            "FillingMaterialResource": {
                "title": "Filling Material Resource",
                "description": "Filling Material resource representation",
                "properties": {
                    "name": {
                        "description": "The name of the filling material",
                        "type": "string"
                    },
                    "color": {
                        "description": "The color of the filling material",
                        "type": "string"
                    },
                    "laboratory": {
                        "$ref": "#/components/schemas/LaboratoryResource"
                    }
                },
                "type": "object"
            },
...
            "LaboratoryResource": {
                "title": "Laboratory Resource",
                "description": "Laboratory resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the laboratory",
                        "type": "string"
                    },
                    "name": {
                        "description": "The name of the laboratory",
                        "type": "string"
                    },
                    "address": {
                        "description": "The address of the laboratory",
                        "type": "string"
                    },
                    "phone": {
                        "description": "The phone number of the laboratory",
                        "type": "string"
                    },
                    "whatsapp": {
                        "description": "The WhatsApp number of the laboratory",
                        "type": "string"
                    },
                    "fillingMaterials": {
                        "description": "List of filling materials associated with the laboratory",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FillingMaterialResource"
                        }
                    }
                },
                "type": "object"
            },
            "MediaResource": {
                "title": "MediaResource",
                "description": "Media resource representation for files and images",
                "required": [
                    "id",
                    "name",
                    "fileName",
                    "collection",
                    "url",
                    "size",
                    "type",
                    "extension",
                    "caption"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the media",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the media file without extension",
                        "type": "string",
                        "example": "nature_photo"
                    },
                    "fileName": {
                        "description": "Complete file name with extension",
                        "type": "string",
                        "example": "nature_photo.jpg"
                    },
                    "collection": {
                        "description": "Collection name where the media is stored",
                        "type": "string",
                        "example": "photos"
                    },
                    "url": {
                        "description": "Full URL to access the media file",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/media/nature_photo.jpg"
                    },
                    "size": {
                        "description": "Human-readable file size",
                        "type": "string",
                        "example": "2.5 MB"
                    },
                    "type": {
                        "description": "Type of media",
                        "type": "string",
                        "enum": [
                            "image",
                            "video",
                            "document",
                            "audio",
                            "other"
                        ],
                        "example": "image"
                    },
                    "extension": {
                        "description": "File extension",
                        "type": "string",
                        "example": "jpg"
                    },
                    "caption": {
                        "description": "Caption or description of the media. Falls back to name if not set",
                        "type": "string",
                        "example": "Beautiful nature photograph"
                    },
                    "width": {
                        "description": "Width of the image in pixels (only for images)",
                        "type": "integer",
                        "format": "int32",
                        "example": 1920,
                        "nullable": true
                    },
                    "height": {
                        "description": "Height of the image in pixels (only for images)",
                        "type": "integer",
                        "format": "int32",
                        "example": 1080,
                        "nullable": true
                    },
                    "thumbnailUrl": {
                        "description": "URL of the thumbnail version. Only present if 'thumb' conversion exists",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/media/conversions/nature_photo-thumb.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MedicalCaseResource": {
                "title": "Medical Case Resource",
                "description": "Medical Case Resource",
                "properties": {
                    "id": {
                        "description": "The ID of the medical case",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the medical case",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the medical case",
                        "type": "string"
                    },
                    "total": {
                        "description": "The total amount associated with the medical case",
                        "type": "number",
                        "format": "float"
                    },
                    "date": {
                        "description": "The date of the medical case",
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/UserResource"
                    }
                },
                "type": "object"
            },
            "MedicalSessionResource": {
                "title": "Medical Session Resource",
                "description": "Medical Session Resource",
                "properties": {
                    "id": {
                        "description": "The ID of the medical session",
                        "type": "integer"
                    },
                    "date": {
                        "description": "The date of the medical session",
                        "type": "string",
                        "format": "date"
                    },
                    "records": {
                        "description": "List of records associated with the medical session",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecordResource"
                        }
                    }
                },
                "type": "object"
            },
...
            "RecordResource": {
                "title": "Record Resource",
                "description": "Record resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the record",
                        "type": "integer"
                    },
                    "type": {
                        "description": "The type of the record",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the record",
                        "type": "string"
                    },
                    "tooth": {
                        "$ref": "#/components/schemas/ToothResource"
                    },
                    "treatment": {
                        "$ref": "#/components/schemas/TreatmentResource"
                    },
                    "fillingMaterial": {
                        "$ref": "#/components/schemas/FillingMaterialResource"
                    },
                    "medicalSession": {
                        "$ref": "#/components/schemas/MedicalSessionResource"
                    },
                    "doctors": {
                        "description": "The doctors associated with the record",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    },
                    "createdAt": {
                        "description": "The creation date of the record",
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "description": "The last update date of the record",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
            "ToothResource": {
                "title": "Tooth Resource",
                "description": "Tooth resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the tooth",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the tooth",
                        "type": "string"
                    },
                    "number": {
                        "description": "The number of the tooth",
                        "type": "integer"
                    },
                    "type": {
                        "description": "The type of the tooth",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TreatmentResource": {
                "title": "Treatment Resource",
                "description": "Treatment resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the treatment",
                        "type": "string",
                        "format": "uuid"
                    },
                    "name": {
                        "description": "The name of the treatment",
                        "type": "string"
                    },
                    "color": {
                        "description": "The color associated with the treatment",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "UserResource",
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the user",
                        "type": "string",
                        "example": 1232132432
                    },
                    "firstName": {
                        "description": "First name of the user",
                        "type": "string",
                        "example": "John"
                    },
                    "lastName": {
                        "description": "Last name of the user",
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "description": "Email address of the user",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "username": {
                        "description": "Username of the user",
                        "type": "string",
                        "example": "john"
                    },
                    "avatar": {
                        "$ref": "#/components/schemas/MediaResource"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}